Move PaymentController to Api/V1 namespace and add PaymentResource for managing payment data
