+++ app/Http/Controllers/Api/V1/PaymentController.php
@@ -1,8 +1,10 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api\V1;
 
+use App\Http\Controllers\Controller;
 use App\Models\Payment;
+use App\Http\Resources\PaymentResource;
 use Illuminate\Http\Request;
 
 class PaymentController extends Controller
